fix search test
