--- lib/content-services/mock/search-control.component.mock.ts
@@ -1,41 +1,0 @@
-/*!
- * @license
- * Copyright 2016 Alfresco Software, Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { Component } from '@angular/core';
-
-@Component({
-    template: `
-        <adf-search-control [highlight]="true">
-                <adf-empty-search-result>
-                    <span id="custom-no-result">{{customMessage}}</span>
-                </adf-empty-search-result>
-        </adf-search-control>
-    `
-  })
-
-  export class SimpleSearchTestCustomEmptyComponent {
-
-    customMessage: string = '';
-
-    constructor() {
-    }
-
-    setCustomMessageForNoResult(message: string) {
-        this.customMessage = message;
-    }
-
-  }
